km-float-field

+++ server/api/mileage/models.py
@@ -6,7 +6,7 @@ from ..users.models import User
 class Mileage(models.Model):
     mileage_id = models.AutoField(primary_key=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    kilometres = models.IntegerField()
+    kilometres = models.FloatField()
     date = models.DateField(default=datetime.date.today)
 
     def __str__(self):
